requirements.txt
python-dotenv==1.0.1
schedule==1.2.1
